requirements.txt
ipykernel
python-dotenv
langchain-community 
langchain 
ollama
langchain-ollama
langchain-core
langgraph
streamlit